backend/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
defusedxml==0.7.1
dj-rest-auth==7.0.1
Django==5.1.6
django-cors-headers==4.7.0
django-rest-auth==0.9.5
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
et_xmlfile==2.0.0
gunicorn==23.0.0
kombu==5.4.2
openpyxl==3.1.3
packaging==25.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pyxform==3.0.1
redis==5.2.1
requests==2.31.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
vine==5.1.0
wcwidth==0.2.13
xlrd==2.0.1
